Delegating providing network type
